DYN-6148 Dynamo Library Lucene Search (#14287)
* DYN-6148 Dynamo Library Lucene Search

Implementing the use of Lucene Search in the Dynamo Library

* DYN-6148 Dynamo Library Lucene Search

Fixing calls to LuceneSearchUtility for using the Lucene Singleton.

* DYN-6148 Dynamo Library Lucene Search

Removing extra space and removing unused usings

* DYN-6148 Dynamo Library Lucene Search Regressions

The SearchNodeTest() test was moved to DynamoLibraryItemsTest due that in the LibraryResourceProviderTests we are not creating the DynamoModel (also there is no way to get the DynamoModel) so the Lucene Singleton is not created and is crashing.
Also I did some minor changes in the test due that the FullName contains the category (so instead of  "dyf://123.456.somepackage.{nodeName}" was replaced by "dyf://Core.Input.{nodeName}" otherwise the Code Block node is not found).
