+++ test/ViewExtensionLibraryTests/DynamoLibraryItemsTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -6,6 +6,13 @@ using System.Reflection;
 using System.Text;
 using Dynamo;
 using Dynamo.Controls;
+using Dynamo.Graph.Nodes.CustomNodes;
+using Dynamo.Interfaces;
+using Dynamo.LibraryViewExtensionWebView2.Handlers;
+using Dynamo.LibraryViewExtensionWebView2;
+using Dynamo.Search.SearchElements;
+using Dynamo.Search;
+using Dynamo.Utilities;
 using Dynamo.ViewModels;
 using Dynamo.Wpf.Interfaces;
 using Moq;
@@ -27,6 +34,53 @@ namespace ViewExtensionLibraryTests
             libraries.Add("DSCoreNodes.dll");
             base.GetLibrariesToPreload(libraries);
         }
+
+        [Test]
+        [Category("UnitTests")]
+        public void SearchNodeTest()
+        {
+            var nodeSearchModel = new NodeSearchModel();
+            var path = @"C:\temp\xyz.dyf";
+            var nodeName = "Code Block";
+            var expectedQualifiedName = $"dyf://Core.Input.{nodeName}";
+
+            for (int i = 0; i < 100; i++)
+            {
+                nodeSearchModel.Add(
+                    new CustomNodeSearchElement(new Mock<ICustomNodeSource>().Object,
+                    new CustomNodeInfo(Guid.NewGuid(), $"Node-{i}", $"Node-{i}-Category{i}", $"Node-{i}-Description", path))
+                    );
+            }
+
+            nodeSearchModel.Add(
+                    new CustomNodeSearchElement(new Mock<ICustomNodeSource>().Object,
+                    new CustomNodeInfo(Guid.NewGuid(), nodeName, "Core.Input", "Node-Description", path))
+                    );
+
+            var pathmanager = new Mock<IPathManager>();
+            var iconProvider = new IconResourceProvider(pathmanager.Object);
+
+            SearchResultDataProvider searchResultDataProvider = new SearchResultDataProvider(nodeSearchModel, iconProvider);
+
+            var extension = string.Empty;
+            var searchResultStream = searchResultDataProvider.GetResource(nodeName, out extension);
+
+            var searchResult = GetLoadedTypesFromJson(searchResultStream);
+            List<LoadedTypeItem> nodesResult = searchResult.loadedTypes;
+
+            Assert.AreEqual(nodesResult.Count, 1);
+            Assert.AreEqual(expectedQualifiedName, nodesResult[0].fullyQualifiedName);
+        }
+
+        private LoadedTypeData<LoadedTypeItem> GetLoadedTypesFromJson(Stream stream)
+        {
+            using (var sr = new StreamReader(stream))
+            {
+                var serializer = new JsonSerializer();
+                return (LoadedTypeData<LoadedTypeItem>)serializer.Deserialize(sr, typeof(LoadedTypeData<LoadedTypeItem>));
+            }
+        }
+
         /*
         [Test]
         [Category("UnitTests"), Category("Failure")]
